fix: Add another condition to the loop check
Technically, not having `is_empty` here will accidentally produce the
right answer *almost all* of the time. This is because there is no way
for an empty board to have a loop in it. However, if the board became
completely dead within `loop_length` turns of initialization, I think we
would get a false positive here.